requirement.txt
streamlit==1.29.0
streamlit-authenticator==0.2.3
mysql-connector-python==8.0.23
pandas>=0.19
sqlalchemy==1.3.19
xlrd==1.2.0
yagmail==0.14.256
PyYAML==6.0
streamlit-aggrid==0.3.4